OK sys_write!

+++ kernel/include/klib/cstr.h
@@ -37,5 +37,5 @@ char *strcpy(char *strDest, const char *strSrc);
 int32_t strncmp(const char* a, const char* b, size_t n);
 char *strncpy(char *dest, const char *src, size_t n);
 char* strcat(char* dest, const char* source);
-uint32_t strlen(const char* str);
+size_t strlen(const char* str);
 int32_t atoi(char *str);
